+++ CHANGES.md
@@ -1,8 +1,8 @@
 CHANGES
 =======
 
-1.2.0.10-SNAPSHOT
------------------
+1.2.0.10
+--------
 
 * Fixed a TypeError in List Selection Should Be. See #18.
 * Fixed a problem with logging a string longer then 1024 characters on Windows systems. See #17.